feat: migrate chat endpoint
